fix: リファクタリングを行い，テストを追加

+++ lib/webapp_log.py
@@ -14,6 +14,7 @@ from webapp_config import APP_URL_PREFIX, LOG_DB_PATH
 from webapp_event import notify_event, EVENT_TYPE
 from flask_util import support_jsonp, gzipped
 import notify_slack
+import traceback
 
 
 class APP_LOG_LEVEL(IntEnum):
@@ -30,9 +31,10 @@ log_lock = None
 log_queue = None
 config = None
 should_terminate = False
+import inspect
 
 
-def init(config_):
+def init(config_, is_read_only=False):
     global config
     global sqlite
     global log_lock
@@ -42,41 +44,59 @@ def init(config_):
 
     config = config_
 
-    sqlite = sqlite3.connect(LOG_DB_PATH, check_same_thread=False)
-    sqlite.execute(
-        "CREATE TABLE IF NOT EXISTS log(id INTEGER primary key autoincrement, date INTEGER, message TEXT)"
-    )
-    sqlite.commit()
-    sqlite.row_factory = lambda c, r: dict(zip([col[0] for col in c.description], r))
+    if sqlite is None:
+        sqlite = sqlite3.connect(LOG_DB_PATH, check_same_thread=False)
+        sqlite.execute(
+            "CREATE TABLE IF NOT EXISTS log(id INTEGER primary key autoincrement, date INTEGER, message TEXT)"
+        )
+        sqlite.commit()
+        sqlite.row_factory = lambda c, r: dict(
+            zip([col[0] for col in c.description], r)
+        )
 
-    should_terminate = False
+    if not is_read_only:
+        should_terminate = False
 
-    log_lock = threading.Lock()
-    log_queue = Queue()
-    log_thread = threading.Thread(target=app_log_worker, args=(log_queue,))
-    log_thread.start()
+        log_lock = threading.Lock()
+        log_queue = Queue()
+        log_thread = threading.Thread(target=app_log_worker, args=(log_queue,))
+        log_thread.start()
 
 
-def term():
+def term(is_read_only=False):
     global sqlite
     global log_thread
     global should_terminate
 
-    should_terminate = True
-    log_thread.join()
-    sqlite.close()
+    if sqlite is not None:
+        sqlite.close()
+        sqlite = None
+
+    if not is_read_only:
+        if log_thread is None:
+            return
+        should_terminate = True
+
+        log_thread.join()
+        log_thread = None
 
 
 def app_log_impl(message, level):
     global config
+    global sqlite
+
     with log_lock:
-        sqlite.execute(
-            'INSERT INTO log VALUES (NULL, DATETIME("now", "localtime"), ?)', [message]
-        )
-        sqlite.execute(
-            'DELETE FROM log WHERE date <= DATETIME("now", "localtime", "-60 days")'
-        )
-        sqlite.commit()
+        if sqlite is None:
+            logging.warning("SQLite is closed.")
+        else:
+            sqlite.execute(
+                'INSERT INTO log VALUES (NULL, DATETIME("now", "localtime"), ?)',
+                [message],
+            )
+            sqlite.execute(
+                'DELETE FROM log WHERE date <= DATETIME("now", "localtime", "-60 days")'
+            )
+            sqlite.commit()
 
         notify_event(EVENT_TYPE.LOG)
 
@@ -90,7 +110,9 @@ def app_log_impl(message, level):
                 config["slack"]["error"]["interval_min"],
             )
 
-        if os.environ.get("DUMMY_MODE", "false") == "true":
+        if (os.environ.get("DUMMY_MODE", "false") == "true") and (
+            os.environ.get("TEST", "false") != "true"
+        ):  # pragma: no cover
             logging.error("This application is terminated because it is in dummy mode.")
             os._exit(-1)
 
@@ -98,18 +120,29 @@ def app_log_impl(message, level):
 def app_log_worker(log_queue):
     global should_terminate
 
+    assert log_queue is not None
+
+    sleep_sec = 0.4
+
     while True:
-        if not log_queue.empty():
-            log = log_queue.get()
-            app_log_impl(log["message"], log["level"])
         if should_terminate:
             break
 
-        time.sleep(0.2)
+        try:
+            if not log_queue.empty():
+                log = log_queue.get()
+                app_log_impl(log["message"], log["level"])
+        except OverflowError:  # pragma: no cover
+            # NOTE: テストする際，freezer 使って日付をいじるとこの例外が発生する
+            logging.debug(traceback.format_exc())
+            pass
+        time.sleep(sleep_sec)
 
 
 def app_log(message, level=APP_LOG_LEVEL.INFO):
     global log_queue
+
+    assert log_queue is not None
 
     if level == APP_LOG_LEVEL.ERROR:
         logging.error(message)
